% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybootASS4.R
\name{myboot}
\alias{myboot}
\title{mybootASS4 (Bootstrap esimators)}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{iter}

\item{n}{n}
}
\value{
histogram for bootstrap sample statistics
}
\description{
mybootASS4 (Bootstrap esimators)
}
