% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2 (Bootstrap esimators)}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{iter}

\item{n}{n}
}
\value{
histogram for bootstrap sample statistics
}
\description{
myboot2 (Bootstrap esimators)
}
