% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{myhyper (Hypergeometric Simultation)}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{iter}

\item{N}{N}

\item{r}{r}

\item{n}{n}
}
\value{
barplot for hypergeometric simultation
}
\description{
myhyper (Hypergeometric Simultation)
}
\examples{
\dontrun{myhyper(100, 20, 12, 5)}
}
