% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newfunctionfromabove.R
\name{newfunctionfromabove}
\alias{newfunctionfromabove}
\title{mycltu (Uniform central limit function)}
\usage{
newfunctionfromabove(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{n}

\item{iter}{iter}
}
\value{
histogram for a uniform central limit function
}
\description{
mycltu (Uniform central limit function)
}
